feat: add error handler for creating group chat and refactor burger button; #23
