+++ react-chat/src/feature/Ripple/Ripple.module.scss
