+++ react-chat/src/feature/Ripple/Ripple.props.ts
