+++ react-chat/src/feature/burgerButton/BurgerButton.tsx
@@ -1,7 +1,7 @@
 import cn from 'classnames';
 import { FC } from 'react';
 
-import { Ripple } from '../Riiple/Ripple';
+import { Ripple } from '../Ripple/Ripple';
 import styles from './BurgerButton.module.scss';
 import { IBurgerButton } from './BurgerButton.props';
 
@@ -11,19 +11,14 @@ export const BurgerButton: FC<IBurgerButton> = ({
   extraClassName,
 }) => {
   return (
-    <button className={cn(styles.button, extraClassName)} onClick={onClick}>
-      <div
-        className={styles.line}
-        data-state={isBurgerActive ? 'open' : 'close'}
-      />
-      <div
-        className={styles.line}
-        data-state={isBurgerActive ? 'open' : 'close'}
-      />
-      <div
-        className={styles.line}
-        data-state={isBurgerActive ? 'open' : 'close'}
-      />
+    <button
+      className={cn(styles.button, extraClassName)}
+      onClick={onClick}
+      data-state={isBurgerActive ? 'open' : 'close'}
+    >
+      <div className={styles.line} />
+      <div className={styles.line} />
+      <div className={styles.line} />
       <Ripple color={'#af5dfc'} duration={2000} />
     </button>
   );
